pyproject.toml
[project]
name = "commstools"
version = "0.1.0"
description = "A Python library for high-performance digital communications research."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "scipy>=1.15.2",
    "sympy>=1.13.3",
    "pydantic>=2.0.0",
    "mypy>=1.19.0",
    "jax[cuda13]>=0.8.1",
    "cupy-cuda13x>=13.6.0",
    "nvidia-curand>=10.4.1.34",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[tool.setuptools]
packages = ["commstools"]
